+++ app/Http/Controllers/Documents/EduSertificateController.php
@@ -7,6 +7,7 @@ use App\Actions\Documents\GenerateEducationSertificatePdf;
 use App\Data\Student\StudentData;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Documents\EduSertificate\EduSertificateIndexFormRequest;
+use App\Services\DocumentService;
 use App\Utils\FileUtils;
 use Carbon\Carbon;
 use Exception;
@@ -15,41 +16,6 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class EduSertificateController extends Controller
 {
-
-    /**
-     * @param array $fileData
-     * TODO как будто бы нужен сервис для этого
-     *
-     * @return array
-     * @throws \Exception
-     */
-    public function getPrepareFileData(array $fileData): array
-    {
-        $prepareFileData = [];
-        foreach ($fileData as $key => $data) {
-            try {
-                $prepareFileData [] = [
-                    'fullname'     => $data[0],
-                    'birthDate'    => Carbon::parse($data[1]),
-                    'program'      => $data[2],
-                    'specialty'    => $data[3],
-                    'course'       => $data[4],
-                    'group'        => $data[5],
-                    'formEdu'      => $data[6],
-                    'formPay'      => $data[7],
-                    'dateStartEdu' => Carbon::parse($data[8]),
-                    'dateEndEdu'   => Carbon::parse($data[9]),
-                    'orderDate'    => Carbon::parse($data[10]),
-                    'orderNum'     => $data[11],
-                    'amount'       => $data[12],
-                ];
-            } catch (\Exception $e) {
-                throw new \Exception('Ошибка в строке ' . $key);
-            }
-        }
-        return $prepareFileData;
-    }
-
     /*
      * Страница формы генерирования справок
      */
@@ -59,14 +25,17 @@ class EduSertificateController extends Controller
         ]);
     }
 
-    public function indexFormRequest(EduSertificateIndexFormRequest $request)
+    /**
+     * @throws Exception
+     */
+    public function indexFormRequest(EduSertificateIndexFormRequest $request, DocumentService $documentService)
     {
         $formData = $request->getData();
         $fileData = FileUtils::readExelFile($formData->file, false);
         if (!$fileData) {
             return back()->withErrors(['Неизвестный тип документа'])->withInput();
         }
-        $prepareFileData = $this->getPrepareFileData($fileData);
+        $prepareFileData = $documentService->getPrepareFileData($fileData);
         $students = StudentData::collection($prepareFileData);
 
         if($formData->docType === 'pdf'){
